requirements.txt
Django==5.1.4
openai==0.28.0
langchain>=0.1.0
langchain-community==0.3.23
openai>=0.28.0
faiss-cpu==1.11.0
tiktoken==0.9.0
python-dotenv>=1.0.0
django-ordered-model==3.7.4
django-admin-interface==0.30.0
requests==2.32.3
gunicorn==23.0.0
psycopg2-binary==2.9.10
environs==14.1.1
whitenoise==6.9.0
